requirements.txt
google-generativeai>=0.3.0
python-dotenv>=0.19.0
selenium>=4.0.0
SpeechRecognition>=3.8.1
PyAudio>=0.2.11
pyttsx3
torch 
openai-whisper
sounddevice 
scipy
